fix make build to generate a binary
This reverts d00b74e2ca660203d4229fb5f1b2316c6028bed0 and 1d865049b39954af5f58d2ea5340e6a9bfdee508, which erroneously removed the `-o` argument from the build command. This was originally done because compilation errors were not detected during the build and a recursive build fixed this. It turns out that the compilation errors were in code that was not called, which is what caused it not to be built.