Fix ticker leaks (#1630)
## Motivation
Tickers should be stopped or it is a leak of its resources.

Closes #1629 

## Changes
Call Stop for each ticker in the system.

## Test Plan
No additional tests.
NumGoRoutines does not include the number of system goroutines. It is possible that a Ticker opens or uses a system goroutine. That's why I couldn't add a test.
If you have ideas on how to easily test it please let me know.