remove leaking metrics in hare (#1650)
## Motivation

these metrics label values grow indefinitely. we want to remove them in order to not blow up our metrics servers. 

Closes #
<!-- `Closes #XXXX, closes #XXXX, ...` links mentioned issues to this PR and automatically closes them when this it's merged -->

## Changes

remove leaking metrics
note: we'll have to revisit this in the future and impl in a way that will use less mem. 
currently we're not using these metrics and they are our heaviest. 
## Test Plan
<!-- Please specify how these changes were tested 
(e.g. unit tests, manual testing, etc.) -->

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed