Perform pingcheck in response to getaddresses
Rather than trying to perform it in response to an incoming ping, which
causes problems. It's better to go ahead and (tentatively) add the peer
(which sent the ping) to the address book, then verify it with a
pingcheck later, when it sends getaddresses. Otherwise we have a race
condition where the getaddresses is received before the pingcheck
occurs.