ATX Builder overhaul (#1619)
## Motivation
Closes #1550 
Closes #1551 
Closes #1552 
Closes #1204 
Closes #1061 
Closes #1494 
Related to #1596, but only partially handles it

## Changes
The node will not start working on an ATX until the previous ATX was included in a block and added to the DB.
Instead of trying to publish an ATX at the start of each layer, ATX creation happens in a loop. When certain required conditions don't apply, execution waits until they do.

The flow is now as follows:
- If a NIPST challenge is stored, but is stale -- clear it.
- If a NIPST challenge is not stored -- create one.
  - Ensure the node is synchronized, or wait if it isn't.
  - Get a positioning ATX.
  - Get the miner's previous ATX.
- Ensure the publication epoch has started, so we can include a full view of the positioning epoch.
- Ensure the node is synchronized.
- Get the view.
- Calculate the active set size from the view.
- Compose, sign and publish the ATX.
- Wait until the ATX is seen in a block.
  - If the ATX's target epoch has passed, wait until the node is synchronized and if it still hasn't been seen in a block consider it an error and start over.
  - If this happens, since the previous ATX is always taken from the database, the node will create a new ATX to replace the one that failed.

## Test Plan
Unit tests were added. All existing tests (unit + automation) pass.