Store state root in processor (#1744)
## Motivation
There was previously no assertion that the state root reported by nodes was not empty (zero), only that it was equal for all nodes. There was also no unit test for `TransactionProcessor.GetStateRoot()`.

This caused an undetected bug to emerge in #1713 where we didn't ever update the `TransactionProcessor.rootHash` and all reported root hashes were empty. @barakshani noticed this in recent runs.

## Changes
Update `TransactionProcessor.rootHash` whenever `addState()` is called.

## Test Plan
- Added an assertion for non-emptiness in `compare_state_roots()`.
- Added a unit test `TestTransactionProcessor_GetStateRoot` that ensures that the state root returned by `GetStateRoot()` matches what was passed to `addState()`.