only close the socket from the higher level. if encountered read/write… (#2141)
… error it's already closed.

## Motivation

Fix the bug that occurred in testnet. 
bug flow: once a connection had to be removed because our cache was full. we would call Close on this connection. which will close the closeChan channel. this will result in an error returned inside `MsgConnection`'s reading flow which will call the same `Close` again in result in close of a closed channel.

## Changes

only close the udp socket from the higher level. if encountered read/write error it's already closed.

## Test Plan
added testing.



Co-authored-by: Lane Rettig <lanerettig@gmail.com>