TestBlockBuilder_BlockIdGeneration
This test depended on the timestamp field to makes blocks different. This is not the correct behavior and now that the timestamp was removed blocks only differed by the coin value (some of the time). This fixes the test by using different ATXs for each block, which makes them consistently different.