Remove CI status badge from README (#2187)
## Motivation
"CI status" for the develop branch is not really well-defined since we don't actually run any CI tasks on the develop branch. Bors runs them on staging, then merges to develop. If we set the status badge to develop, as it is now, it shows "no status." If we set it to staging, it'll show the status of the last bors run, which may have failed (and not been merged), which is misleading.

@noamnelke you were right, I concede :)