Only report receiving a block once (#2437)
We've seen cases in system tests where nodes report getting several multiples of the number of blocks that are supposed to be in a layer. It appears to happen because we log receiving a block before checking if we know it already, at which point block processing stops.

Log `got new block` only after checking for duplicates.

Since this only affects logging, no new tests are needed.

<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)