Get a meaningful branch name in detached mode (#2392)
## Motivation
When running tests locally in detached mode, the branch gets set to `HEAD` and the docker image is always pushed to `spacemeshos/go-spacemesh-dev:HEAD`. This means that multiple tests running on different code can overwrite each other. In this case, use the commithash instead.

## Changes
Use commithash instead of `HEAD` as branch for docker when running in detached mode.

## Test Plan
N/A

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [ ] Test changes and document test plan
- [ ] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)