Remove circular dependency in Hare active set/eligibility (#2357)
## Motivation
Fixes #2356

## Changes
When calculating active set for a given epoch for hare eligibility, use the Tortoise active set rather than Hare active set. Hare active set consists of the set of unique node IDs that produced contextually valid blocks for a "safe" layer in the past. The problem is that this causes a circular dependency: if Hare fails for the "safe" layer, then the active set size for a future layer will be zero, causing the beacon to fail and eligibility to always be false. The Tortoise active set, by contrast, consists of the set of unique node IDs that produced valid ATXs targeting the epoch for the layer. This is much more robust against Hare failure and removes the circular dependency.

## Test Plan
Updates existing tests accordingly

## TODO
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [ ] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)