remove MessageType duplication (#2515)
## Motivation
current code the server.MessageType definition reside in the caller. this is error-prone since each client defines different types to the same constant. currently each caller create its instance of MessageServer so this is not a bug now.

move the definition to the service side has the following benefit
- avoid future coding error where multiple handler is registered to the same constant
- ease of debugging, since each MessageType 

same goes for fetch.Hint

## Changes
move MessageType definition to msgserver.go and Hint to fetch.go

## Test Plan
make test

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)