mesh: fix origin/destination index key to avoid unintentional range queries (#2536)
## Motivation
closes: https://github.com/spacemeshos/go-spacemesh/issues/2530

## Changes
- changed keys encoding for origin/destination transaction index to use big endian instead of strconv
if there are multiple keys with '2' in it, such as '2', '20', '200'. utils.BytePrefix (that is used in datatabase iterator) won't terminate on key '2', but instead scan all of them. For '20' it will scan '20', '200' and so on.

I changed it to big endian specifically (and not little endian) because in future we may use it for range queries instead of iterating over all layers and making point queries.

## Test Plan
unit tests

## TODO
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)