relax sync failure on ATX failure (#2571)
## Motivation
when syncing current epoch, sync fails whenever it fails to fetch an ATX, and it will take the node from synced state to notSynced state. this is possibly too aggressive for current epoch, as ATXs are not fully gossiped in the network.

## Changes
for current epoch, only fail sync on ATX failure at last layer. 
for previous epoch, sync fails as soon as it fails to fetch ATXs for that epoch.

## Test Plan
unit test

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)