Don't hold mutex while sending on connection (#2435)
## Motivation
Partially addresses #1868, #2386 (and maybe #2405)

## Changes
The underlying connection objects contain a mutex that's used to protect the `closed` struct member. They hold the mutex while trying to send on the socket. But the send operation may be delayed or time out, which has a ripple effect up the stack: it blocks the operation that enqueues outbound messages, which in turns blocks propagation of gossip messages (#1868) (since gossip waits for outbound messages to all peers to be enqueued before it processes the next outbound gossip message). There are various places in this stack where we could add more concurrency or queues, but ultimately this appears to be the bottleneck, and fixing it should address the gossip message bottleneck.

## Test Plan
Adds a test that makes sure that a blocked Write operation does not block Send

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [ ] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)