p2p node: Package improvements (#2851)
## Motivation
As stated in Discord, I have been reviewing the whole project and hope to open a series of PRs that clean up a few things in the codebase and make things more idiomatic in general. There's a ton of good Go code here and my hope is that these changes can only make it better, more dependable, and idiomatic.

Instead of opening a huge PR that spans several packages, I figured it would be best to open smaller PRs that are package-focused. I have started with the p2p package and have several PRs to come.

## Changes
- Remove unneeded pkg var: `emptyNode`
- Clean up resource leak potentials with open file handlers
- Add more context to errors
- Idiomatic variable naming

## Test Plan
Unit tests all pass

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [ ] Test changes and document test plan
- [ ] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)