sync: error if some requested blocks were not fetched or requested layer is not yet processed (#2853)
## Motivation

These were two edge cases that were causing an observed issue with missing blocks.

If blocks were requested from the first peer then all other requests will automatically be successful. So even if the initial request will complete only partially the layer won't be failed and will be sent for the verification to the tortoise.

Sometimes the only successful response is from the peer that is lagging behind, or may have only a partial state (e.g. requested layer is 101, processed layer in response is 100). In this case, we can treat this response as failed instead of processing it. Also instead of failing when the node was already synced, we will simply interrupt the sync earlier.

I think we can also recover by fetching base blocks (and not only exceptions), but i wasn't able to test base block change on current devnet.

Closes https://github.com/spacemeshos/go-spacemesh/issues/2847

## Changes
- track blocks that were requested but not received and fail layer if we are not able to fetch all blocks
- if the response has a processed layer lower than the requested layer mark it as failed
- fetch base block in handle block data

## Test Plan
<!-- Please specify how these changes were tested 
(e.g. unit tests, manual testing, etc.) -->

## TODO
<!-- This section should be removed when all items are complete -->
- [ ] Explain motivation or link existing issue(s)
- [ ] Test changes and document test plan
- [ ] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)