Wrap most errors with additional context (#2855)
## Motivation
<!-- Please mention the issue fixed by this PR or detailed motivation -->
<!-- Closes # -->
<!-- `Closes #XXXX, closes #XXXX, ...` links mentioned issues to this PR and automatically closes them when this it's merged -->
Currently, many error logs are not descriptive enough. E.g., there're many errors like `$ERROR: leveldb: not found`, where it's not clear what's the origin of the error. Also, it's a good practice to wrap errors is Go. 

The PR wraps most of the errors with additional context. Most of them were found with the `wrapcheck` linter, the other part manually. There are still some unwrapped errors inside packages, but I'm not sure we need to wrap them. 

There also will be another PR, which adds the `wrapcheck` linter to the CI, so that we don't add new unwrapped errors.

## Changes
<!-- Please describe in detail the changes made -->
- Wrap most errors with additional context

## Test Plan
<!-- Please specify how these changes were tested 
(e.g. unit tests, manual testing, etc.) -->
- Unit tests
- System tests

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)