tortoise: use expected layer weight when computing thresholds (#2963)
## Motivation

This is basic for security of the tortoise, otherwise voting is disproportionate to the actual weight and block can be verified even when they supported by true majority (see TestVerifyLayerByWeightNotSize) . 

Also important for recovery from the partition where we learn about new atxs. During rerun we will see that the weight increases but we can't make progress with votes from a single beacon, so the verified layer will be stuck until votes from ballots with "bad" beacon can be counted. Will send a followup change for this use case. 

closes https://github.com/spacemeshos/go-spacemesh/issues/2946

## Changes
- expected weight of votes for a block in layer N is derived from atxs of the ballots from all layers after N up to last known

## Test Plan
TestVerifyLayerByWeightNotSize, TestComputeExpectedWeight and other uts

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)