Deprecate usage of `svm/state` and replace it with `svm` at calling sites  (#2992)
## Motivation
We're hard deprecating the `svm/state` package, which will remain in the `go-spacemesh` as a temporary solution before we make the switch to `go-svm`. This task is tracked in https://github.com/spacemeshos/go-spacemesh/issues/2828.

## Changes
Making sure that `svm/state` is not used anywhere in the codebase and that all calling sites interact directly with the `svm` package instead.

## Test Plan
No functional changes are expected. Some tests, however, require changes in how they interact with `svm` as well.

## DevOps Notes
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)


Co-authored-by: Filippo Costa <filippocosta.italy@gmail.com>