tortoise: rational numbers for weight (#2996)
## Motivation

I intended to replace big.Float with fixed float library. But realized that rational number's make more sense here, we are using two distinct weights:
- counted weight from ballots, each unit is atx_weight / eligibility_count
- expected weight from epochs in a single, each unit is total_atx_weight / layers_in_epoch  
Later expected weight is multiplied by a fraction, that is also represented as a rational number. And after multiplication two weights are compared. 

So it seems just more straightforward to use rational numbers in this case. They are slightly less performant for full tortoise, but it doesn't seem to matter.

## Changes
- use big.Rat instead of big.Float for weight

## Test Plan
ut's

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)