tortoise: depreoritize ballots with bad beacons (#3011)
## Motivation

if tortoise is in verifying mode it is using last good ballot as a base ballot, and encodes only votes starting from the base ballot layer. once switched into full mode we can't rely on this optimization anymore and need to recompute the least bad ballot, and encode votes for every layer that we have in memory. in such case we shouldn't select base ballot with bad beacon, as it complicates procedure for switching tortoise into verifying mode.

reproduced this during testing.

## Changes
- use ballots with a bad beacon only if there are no ballots with good beacon

## Test Plan
unit tests

## TODO
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)