tortoise: recompute verification window when switching into full mode (#3041)
## Motivation

verifying has a large verification window (think 1_000_000) and if it failed to verify layer the threshold will be computed according to that window. if we won't reset the threshold full tortoise will have to count votes for 1_000_000 layers before
any layer can be expected to get verified. this is infeasible given current performance of the full tortoise and may take weeks to finish.

closes https://github.com/spacemeshos/go-spacemesh/issues/3009

## Changes
- compute thresholds according to the full tortoise configuration when in full mode

## Test Plan
TestRerunDistanceVoteCounting