tortoise: encode abstain votes as an array with layers (#3043)
## Motivation

In verifying tortoise we should not count weight from the ballot if the ballot abstained for that layer. This will be easier to implement if abstain votes were encoded as an array of layer. 

Also, this change is a must for UBC, as we don't know the actual block value before hare termination. So there is no way to vote abstain.

Support/Against are not collapsed into a single list as it leads to less efficient implementation. Maybe we should revise it in the future.

Closes: https://github.com/spacemeshos/go-spacemesh/issues/2369


## Changes
- encode abstain votes as an array of layers

## Test Plan
existing unit tests, logically nothing has changed