mesh: prefixes in the same space should not overlap (#3051)
## Motivation

prefixes in the same space should not overlap. as it leads to corrupted database state.
for example keys `bl || abc` and `b || lzxd` overlap, so when we query for `bl` we could unexpectedly load object with key zxd.

## Changes
- use nonoverlapping prefixes for blocks and ballots

## Test Plan
i don't know how to add a failing test if prefixes are correct, so i added an example that would fail if prefixes are incorrect in a specific way.