tortoise: add OnBlock and OnBallot to tortoise interface (#3053)
## Motivation

Several bugs can occur because tortoise may miss state that wasn't available at the time when the layer was passed to the tortoise.

A ballot that uses a base ballot that wasn't available when layer was created will not be marked good and will implicitly vote for many blocks in the past. This problem will be fixed during rerun, but we can avoid it.

If block was received after layer was processed we won't count votes for it and can verify layer without deciding on that particular block. This again will be fixed by rerun, but it can be avoided as well.

Besides that OnBallot will be required for work on this issue https://github.com/spacemeshos/go-spacemesh/issues/3054 .

OnBlock will be removed after pruning is integrated, as votes will have all information that we need in tortoise (layer id and block id). 

closes: https://github.com/spacemeshos/go-spacemesh/issues/3004

## Changes
- introduce OnBlock and OnBallot methods for tortoise
- call them from mesh

## Test Plan
unit tests