sql: store ballots, blocks and layers in sqlite  (#3047)
## Motivation

The existing approach lacks atomicity, causal durability (e.g. an atx may not be on disk when a ballot is saved), and the durability can't be enforced in general without running every leveldb operation in sync mode. All this problems will result in subtle bugs that are hard to diagnose.

For those 3 requirements, we want to maintain all state in a single db. Moving state to a single leveldb will require us to enforce isolation ourselves (by maintaining separate namespace and manually concatenating keys like we do in some modules), beside that we have to create every single index manually while with sqlite we can just do `CREATE INDEX` and sqlite will do it for us and probably do a better job. 

Another significant benefit is that we can duplicate some state in sql table to avoid loading the whole structure into memory. For instance it will be relevant for atx, which is a large (10kb) and usually, after it was validated, we want to know only the associated smesher and weight of the atx. This would be problematic with leveldb, and would require adding custom index.

related: https://github.com/spacemeshos/go-spacemesh/issues/2918

## Changes
- general plumbing for core database stuff (db, transaction, migrations) using https://github.com/crawshaw/sqlite that is a relatively simple wrapper around C sqlite 
- tables and for layers, blocks and ballots
- reworked ZeroLayer, it is relevant only for hare_output, which can be in 3 states - nil, empty, non-empty. 
  SetZeroLayer update hare_output to empty state, at which tortoise will vote against all blocks within hdist. 
- updated to golang 1.16 for `embed` module. note that there is a bug with go mode tidy in 1.16 so i had to manually fix go.sum and disable go mod tidy on ci - https://github.com/golang/go/issues/44129

## Test Plan
existing and new uts