beacon: a little smaller vote bit vector (#3086)
## Motivation
<!-- Please mention the issue fixed by this PR or detailed motivation -->
Part of #3072
<!-- `Closes #XXXX, closes #XXXX, ...` links mentioned issues to this PR and automatically closes them when this it's merged -->

## Changes
<!-- Please describe in detail the changes made -->
- collapse valid and potentially valid votes for codec. we don't need this distinction anymore once we cast votes for the first voting round.

- use big.Int to encode bit vector instead of using extra lib bits-and-blooms/bitset
this also reduces the resolution of bit vector from 8 bytes to 1 byte.
  - before
it takes 16 bytes (2 uint64) to encode votes for 100 proposals and 4 bytes for length => 20 bytes total
it takes 8 bytes (1 uint64) to encode votes for 42 proposals and 4 bytes for length => 12 bytes total

  - after
this will encode the votes to as many bytes as needed. 
it takes 13 bytes (CEIL(100 / 8)) to encode votes for 100 proposals and 4 bytes for length => 17 bytes total
it takes 6 bytes (CEIL(42 / 8)) to encode votes for 42 proposals and 4 bytes for length => 10 bytes total

100 is the vote limit
42 is the expected number of votes

## Test Plan
<!-- Please specify how these changes were tested 
(e.g. unit tests, manual testing, etc.) -->
unit test

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)