conservative cache and mempool iterator (#3152)
## Motivation
<!-- Please mention the issue fixed by this PR or detailed motivation -->
Closes #3034
Partial implementation of #3094
<!-- `Closes #XXXX, closes #XXXX, ...` links mentioned issues to this PR and automatically closes them when this it's merged -->

## Changes
<!-- Please describe in detail the changes made -->
- mempool iterator spec'ed in #3034 
- add conservative cache spec'ed in #3094
a notable difference btwn the spec and the implementation is the lack of the layer index in the cache.
my first draft of implementation organized the data structure as spec'ed
```
principle
 ↳ layer
   ↳ nonce
     ↳ (transaction_id, c_amount, c_fee)
```
however, due to the fact that layer is an artificial property of the account data (used when packing a new proposal) and that  ultimately transactions need to be applied in nonce order, the extra layer index becomes redundant and cumbersome. this is because even if we gather all the transactions marked as the layer to apply, one still needs to check if the transactions are applied in nonce order.

moreover, the layer information is used onl when collecting mempool transactions into a proposal. i.e. a tx with an empty layer ID is eligible for the next proposal. it does not really matter which layer it is in, as long it is after the last applied layer. 

the implementation now uses layer as the property of each nonce. 
```
principle
   ↳ nonce (with layer as property)
     ↳ (transaction_id, c_amount, c_fee)
```
that way, we simply apply transactions in nonce order, and error out otherwise. 
when packing the next proposal, we find the first nonce that doesn't have layer attached, and pack the proposal from that nonce and up from the cache.

this is the first PR of the series. the next PR will do the integration with mesh and the database.

## Test Plan
<!-- Please specify how these changes were tested 
(e.g. unit tests, manual testing, etc.) -->
unit tests

## TODO
<!-- This section should be removed when all items are complete -->
- [ ] Test changes and document test plan

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)